Bug fix, game un-runnable
